bing_crawler: raise ParseError
Currently, if we do no find any CSS classes matching the ones we expect we
print a ParseError. This is a critical error as it means we are not able to
actually get any salient information from the page we are currently trying to
scrape and thus should abort abruptly.

This commit raises the ParseError and catches it up the stack. All off the
crawlers error catching/handling needs more thought and is currently not done
particularly well so we will better handle this error in the future during that
rework. For now it is good enough that the error is raised and not printed.