CI: only run push events to master branch
Currently, we run all CI checks on every branch however, this is redundant
as we also run CI checks during a PR meaning when a PR is opened all checks run
twice.

This commit ensures that all CI checks are ran fully when merged to master as
the PR CI check will ensure that checks pass at branch level.