crawl/_api.py: ensure any spaces betweent metadata has dashes
Currently, if a course name, level or module contains any spaces it gets
translated as is when the path is created to save the topic list items.
This is a pain to traverse because spaces need special character to escape them.
Furthermore, it could lead to issues in some file systems that may not allow
spaces in directory names.

This commit ensures that the meta data field in the search results object
replaces any spaces with dashes.