syllime-crawl@.service: never restart systemd service
Currently, the systemd service always restarts regardless of the exit status
of the crawler process. This raises serval problems:
- if the process exists successfully we dont need to recrawl everything
- if the process exists unsuccessfully that means there was an unexpected error
and we need to handle this properly e.g. if we get blocked by a website or face
a CAPTCHA. We dont want to keep bombing the site with requests.

This commit changes the restart policy of the systemd service to never restart,
currently we want starting the systemd service to be done manually.