sylli_crawl->driver_manager.py: raise WebDriverException
This commit raises the WebDriverException in the driver code rather than
letting the program silently carry on and trying to catch a RunTimeError
later on.

This commit also removes the try/except for the RunTimeError as this seems
like an anti-pattern. We'll let the error bubble up to the `fetch` call on the
class and deal with it there.