crawler/_api.py: fix typo
This commit fixes a typo that has been causing the process to fail in prod.

PDF pages were not correctly getting picked up and thus being crawled like
regular html pages, causing the runner to fail as a PDF in the browser is
essentially a picture loaded into an iframe and not a regular html page.

This commit fixed the issues and adds test cases for all paths the crawler could
take so we can pick up on this issue in the future.